Evict large function definitions from the Helpers.hh header file
